Monster Name Crash Fix
Overwriting Monster Names was safe without modifying the return value for Combinations, but not for viewing Monster Card data. It accesses the information immediately preceding the proper name list (reading and writing to the -1 index).

This now alters things to replace Pixie's data when getting an invalid entry, and write it back if we are checking for Pixie. Simple fix.