Added TechsRaw
Not big brained enough at the moment to do the mapping between the TechSlots and the TechIDs that need to be used here. If we had the techs it would be easy but they aren't here so then you need to check the table to figure out which Slot=ID and then set the bits appropriately. This is a hacky fix that basically makes TechSlots for show but at the same time right now TechSlots doesn't set the correct bits.

I loaded a new save, a color pandora should have had Tail Swing, Face Attack ,and MegaShotgun according to the debug output (M0, C0, L0).
What they actually had was Face Attack, Cracker, and nothing else. This maps to the tech ID mapping of 0, 6, and 18 (0 is FA, 6 is Cracker, which is actually M0, and ColorP tech IDs end at 15 as it only has 16 techs total).